docs/integration/actions/MCP/redshift-data-query-mcp/pyproject.toml
[project]
name = "redshift-data-query-mcp"
version = "0.1.0"
description = "Amazon Redshift Data Query MCP integration with QuickSuite"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
    "aws-cdk.aws-lambda-python-alpha",
]
